+++ pkg/action/metrics.go
@@ -15,21 +15,23 @@ var (
 )
 
 var (
-	requestDuration = prometheus.NewHistogram(
+	requestDuration = prometheus.NewHistogramVec(
 		prometheus.HistogramOpts{
 			Namespace: namespace,
 			Name:      "request_duration_seconds",
 			Help:      "Histogram of latencies for requests to the Hetzner API.",
 			Buckets:   []float64{0.001, 0.01, 0.1, 0.5, 1.0, 2.0, 5.0, 10.0},
 		},
+		[]string{"project"},
 	)
 
-	requestFailures = prometheus.NewCounter(
+	requestFailures = prometheus.NewCounterVec(
 		prometheus.CounterOpts{
 			Namespace: namespace,
 			Name:      "request_failures_total",
 			Help:      "Total number of failed requests to the Hetzner API.",
 		},
+		[]string{"project"},
 	)
 )