+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Switch to cloud.drone.io for CI
 
+### Added
+
+* Support multiple credentials and config file
+
 ## [0.2.0] - 2019-01-11
 
 ### Added